remove hooks; move request matching to request class
